+++ src/components/explore/ExploreItems.jsx
@@ -1,78 +1,71 @@
-import React from "react";
-import { Link } from "react-router-dom";
-import AuthorImage from "../../images/author_thumbnail.jpg";
-import nftImage from "../../images/nftImage.jpg";
+import axios from "axios";
+import React, { useEffect, useState } from "react";
+import Item from "../UI/Item"
 
 const ExploreItems = () => {
+  const [list, setList] = useState([])
+  const [displayedList, setDisplayedList] = useState([])
+  const [loading, setLoading] = useState(false)
+  const [filter, setFilter] = useState("")
+  const [itemsToShow, setItemsToShow] = useState(8)
+
+  useEffect(() => {
+    const getList = async () => {
+      try {
+        setLoading(true);
+        const apiUrl = `https://us-central1-nft-cloud-functions.cloudfunctions.net/explore?filter=${filter || "likes_high_to_low"}`
+        const response = await axios.get(apiUrl)
+        setList(response.data)
+        setDisplayedList(response.data.slice(0, 8))
+        setItemsToShow(8)
+      } catch (error) {
+        console.error("Error fetching Explore page:", error)
+      } finally {
+        setLoading(false)
+      }
+    }
+
+    getList()
+  }, [filter])
+
+  const loadMore = () => {
+    const newItemsToShow = itemsToShow + 4;
+    setDisplayedList(list.slice(0, newItemsToShow));
+    setItemsToShow(newItemsToShow);
+  };
+
   return (
     <>
       <div>
-        <select id="filter-items" defaultValue="">
+        <select id="filter-items" value={filter} onChange={(e) => setFilter(e.target.value)}>
           <option value="">Default</option>
           <option value="price_low_to_high">Price, Low to High</option>
           <option value="price_high_to_low">Price, High to Low</option>
           <option value="likes_high_to_low">Most liked</option>
         </select>
       </div>
-      {new Array(8).fill(0).map((_, index) => (
-        <div
-          key={index}
-          className="d-item col-lg-3 col-md-6 col-sm-6 col-xs-12"
-          style={{ display: "block", backgroundSize: "cover" }}
-        >
-          <div className="nft__item">
-            <div className="author_list_pp">
-              <Link
-                to="/author"
-                data-bs-toggle="tooltip"
-                data-bs-placement="top"
-              >
-                <img className="lazy" src={AuthorImage} alt="" />
-                <i className="fa fa-check"></i>
-              </Link>
-            </div>
-            <div className="de_countdown">5h 30m 32s</div>
 
-            <div className="nft__item_wrap">
-              <div className="nft__item_extra">
-                <div className="nft__item_buttons">
-                  <button>Buy Now</button>
-                  <div className="nft__item_share">
-                    <h4>Share</h4>
-                    <a href="" target="_blank" rel="noreferrer">
-                      <i className="fa fa-facebook fa-lg"></i>
-                    </a>
-                    <a href="" target="_blank" rel="noreferrer">
-                      <i className="fa fa-twitter fa-lg"></i>
-                    </a>
-                    <a href="">
-                      <i className="fa fa-envelope fa-lg"></i>
-                    </a>
-                  </div>
-                </div>
-              </div>
-              <Link to="/item-details">
-                <img src={nftImage} className="lazy nft__item_preview" alt="" />
-              </Link>
-            </div>
-            <div className="nft__item_info">
-              <Link to="/item-details">
-                <h4>Pinky Ocean</h4>
-              </Link>
-              <div className="nft__item_price">1.74 ETH</div>
-              <div className="nft__item_like">
-                <i className="fa fa-heart"></i>
-                <span>69</span>
-              </div>
-            </div>
+      {loading ? (
+        <p>Loading...</p>
+      ) : (
+        displayedList.map((item) => (
+          <div
+            key={item.nftId}
+            className="d-item col-lg-3 col-md-6 col-sm-6 col-xs-12"
+            style={{ display: "block", backgroundSize: "cover" }}
+          >
+            <Item item={item} />
           </div>
+        ))
+      )}
+
+      {itemsToShow < list.length && (
+        <div className="col-md-12 text-center">
+          <button type="button" onClick={loadMore} className="btn-main lead">
+            Load more
+          </button>
         </div>
-      ))}
-      <div className="col-md-12 text-center">
-        <Link to="" id="loadmore" className="btn-main lead">
-          Load more
-        </Link>
-      </div>
+      )}
     </>
   );
 };
